13/11/2017 - 16:58 -> Upload debug at ft_strcpy

+++ ft_strnequ.c
@@ -6,7 +6,7 @@
 /*   By: mabessir <mabessir@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/10 11:20:36 by mabessir          #+#    #+#             */
-/*   Updated: 2017/11/10 15:31:55 by mabessir         ###   ########.fr       */
+/*   Updated: 2017/11/13 12:25:11 by mabessir         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int		ft_strnequ(char const *s1, char const *s2, size_t n)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (s1[i] || i < n)
