+++ ft_strsub.c
@@ -6,7 +6,7 @@
 /*   By: mabessir <mabessir@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/10 11:24:14 by mabessir          #+#    #+#             */
-/*   Updated: 2017/11/10 15:37:47 by mabessir         ###   ########.fr       */
+/*   Updated: 2017/11/13 12:47:52 by mabessir         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ char	*ft_strsub(char const *s, unsigned int start, size_t len)
 	str = (char *)malloc(sizeof(str) * (len));
 	if (!str)
 		return (NULL);
-	while (len >= 0)
+	while ((int)len >= 0)
 	{
 		str[i] = ((char *)s)[start];
 		start++;
